feature:  Add min app version server
