+++ core/common/src/main/java/com/teamoffroad/core/common/data/di/RepositoryModule.kt
@@ -1,8 +1,10 @@
 package com.teamoffroad.core.common.data.di
 
 import com.teamoffroad.core.common.data.repository.AutoSignInRepositoryImpl
+import com.teamoffroad.core.common.data.repository.MinSupportedVersionRepositoryImpl
 import com.teamoffroad.core.common.data.repository.TokenRepositoryImpl
 import com.teamoffroad.core.common.domain.repository.AutoSignInRepository
+import com.teamoffroad.core.common.domain.repository.MinSupportedVersionRepository
 import com.teamoffroad.core.common.domain.repository.TokenRepository
 import dagger.Binds
 import dagger.Module
@@ -24,4 +26,11 @@ abstract class RepositoryModule {
     abstract fun bindAutoSignInRepository(
         authRepositoryImpl: AutoSignInRepositoryImpl,
     ): AutoSignInRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindMinSupportedVersionRepository(
+        minSupportedVersionRepositoryImpl: MinSupportedVersionRepositoryImpl
+    ): MinSupportedVersionRepository
+
 }
