+++ core/common/src/main/java/com/teamoffroad/core/common/domain/model/MinSupportedVersion.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.main.domain.model
+package com.teamoffroad.core.common.domain.model
 
 data class MinSupportedVersion(
     val ios: String,
