+++ core/common/src/main/java/com/teamoffroad/core/common/domain/repository/MinSupportedVersionRepository.kt
@@ -1,0 +1,8 @@
+package com.teamoffroad.core.common.domain.repository
+
+import com.teamoffroad.core.common.domain.model.MinSupportedVersion
+
+
+interface MinSupportedVersionRepository {
+    suspend fun fetchMinSupportedVersion(): MinSupportedVersion
+}
