--- feature/main/src/main/java/com/teamoffroad/feature/main/data/di/NetworkModule.kt
@@ -1,19 +1,0 @@
-import com.teamoffroad.feature.main.data.remote.service.MinSupportedVersionService
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-
-    @Provides
-    @Singleton
-    fun provideMinSupportedVersionService(retrofit: Retrofit): MinSupportedVersionService {
-        return retrofit.create(MinSupportedVersionService::class.java)
-    }
-}
-
