--- feature/main/src/main/java/com/teamoffroad/feature/main/domain/repository/MinSupportedVersionRepository.kt
@@ -1,7 +1,0 @@
-package com.teamoffroad.feature.main.domain.repository
-
-import com.teamoffroad.feature.main.domain.model.MinSupportedVersion
-
-interface MinSupportedVersionRepository {
-    suspend fun fetchMinSupportedVersion(): MinSupportedVersion
-}
